# gin-front-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博客用户后台API接口文档",
        "title": "博客用户后台",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/bArticle/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增文章",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章相关接口"
                ],
                "summary": "新增文章",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BArticleDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/bArticle/articleId": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据id查询文章",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章相关接口"
                ],
                "summary": "根据id查询文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/bArticle/detail": {
            "get": {
                "description": "查询文章详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章相关接口"
                ],
                "summary": "查询文章详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/bArticle/keyword/list": {
            "get": {
                "description": "根据关键字分页查询文章列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章相关接口"
                ],
                "summary": "根据关键字分页查询文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页数",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文章标题或用户名称",
                        "name": "keywords",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/bArticle/like": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "文章点赞和取消",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章相关接口"
                ],
                "summary": "文章点赞和取消",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IdDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/bArticle/list": {
            "get": {
                "description": "分页查询文章列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章相关接口"
                ],
                "summary": "分页查询文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页数",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签id",
                        "name": "tagId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类id",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "查询参数：1-\u003e热榜,2-\u003e发布,3-\u003e最新",
                        "name": "orderType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BArticleDto": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "description": "分类id",
                    "type": "integer"
                },
                "content": {
                    "description": "文章内容",
                    "type": "string"
                },
                "id": {
                    "description": "文章id",
                    "type": "integer"
                },
                "image": {
                    "description": "封面地址",
                    "type": "string"
                },
                "markdownContent": {
                    "description": "文章内容md版",
                    "type": "string"
                },
                "summary": {
                    "description": "文章简介",
                    "type": "string"
                },
                "tagId": {
                    "description": "标签id",
                    "type": "integer"
                },
                "title": {
                    "description": "文章标题",
                    "type": "string"
                }
            }
        },
...
        "model.IdDto": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID",
                    "type": "integer"
                }
            }
        },
...
        "result.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回的数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}